Add unit test for genvi command line interface
